+++ Task_2_1_1/src/main/java/ru/nsu/kuklin/parallelprime/CompositeNumberDetector.java
@@ -1,96 +1,13 @@
 package ru.nsu.kuklin.parallelprime;
 
-import java.util.Arrays;
-import java.lang.ThreadGroup;
-import java.lang.Thread;
-import java.util.concurrent.atomic.AtomicBoolean;
-
 /**
- * Class to detect composite numbers in arrays.
+ * Interface for Detectors of composite numbers.
  */
-public class CompositeNumberDetector {
+public interface CompositeNumberDetector {
     /**
-     * Checks if there is a composite number in array.
-     * Sequential
+     * Main method.
+     * @param nums To search a composite number in
+     * @return Whether there are composite numbers inside
      */
-    public static boolean detectSequential(int[] nums) {
-        for (int i = 0; i < nums.length; i++) {
-            i++;
-            boolean prime = true;
-            for (int j = 2; j * j < nums[i]; j++) {
-                if (nums[i] % j == 0) {
-                    prime = false;
-                    break;
-                }
-            }
-            if (!prime) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Checks if there is a composite number in array.
-     * Uses parallel stream
-     */
-    public static boolean detectParallelStream(int[] nums) {
-        return Arrays
-            .stream(nums)
-            .parallel()
-            .map(num -> {
-                for (int j = 2; j * j < num; j++) {
-                    if (num % j == 0) {
-                        return 0;
-                    }
-                }
-                return 1;
-            })
-            .anyMatch(p -> p == 0);
-    }
-
-    /**
-     * Checks if there is a composite number in array.
-     * Uses raw threads
-     */
-    public static boolean detectParallelThreads(int[] nums, int threadCount) {
-        ThreadGroup group = new ThreadGroup("Detectors");
-        AtomicBoolean result = new AtomicBoolean(false);
-        int next = 0;
-        for (int i = 0; i < threadCount && next < nums.length; i++) {
-            int l = next;                               // Equally distribute surplas numbers
-            int r = l + nums.length / threadCount + (i < nums.length % threadCount ? 1 : 0);
-            Thread t = new Thread(group, new Runnable() {
-                public void run() {
-                    for (int k = l; k < r; k++) {
-                        for (int j = 2; j * j < nums[k]; j++) {
-                            if (nums[k] % j == 0) {
-                                result.set(true);
-                                group.interrupt();
-                                break;
-                            }
-                        }
-                        if (Thread.currentThread().isInterrupted()) {
-                            return;
-                        }
-                    }
-                }
-            });
-            t.start();
-            next = r;
-        }
-        
-        Thread[] threads = new Thread[threadCount];
-        int cnt = group.enumerate(threads);
-        for (int i = 0; i < cnt; i++) {
-            try {
-                threads[i].join();
-            } catch (InterruptedException e) {
-                System.out.println("Interrupted");
-            }
-        }
-
-        return result.get();
-    }
+    boolean detect(int[] nums);
 }
-
